+++ test/typegen/client/index.ts
@@ -1,2 +1,2 @@
-export { client as testLayoutLayout } from "./testLayout";
-export { client as weirdPortalsLayout } from "./weirdPortals";
+  export { client as testLayoutLayout } from "./testLayout";
+  export { client as weirdPortalsLayout } from "./weirdPortals";